+++ CHANGELOG.md
@@ -25,6 +25,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2022-06-01
+
+### Added
+
+- Added a new entry point which just displays the help information for the options
+  which will be exposed as options in a product's installer script that updates
+  CFS configurations.
+- Added a new entry point which pre-processes the options passed to the main
+  entry point to determine which options refer to files and then outputs the
+  mount options that should be passed to the container runtime and the translated
+  arguments that refer to file paths as mounted inside the container.
+
 ## [3.0.1] - 2022-05-25
 
 ### Changed